+++ CHANGELOG.md
@@ -12,6 +12,8 @@ about writing changes to this log.
 
 - [TVIST1-743](https://jira.itkdev.dk/browse/TVIST1-743)
   Added case events.
+- [TVIST1-788](https://jira.itkdev.dk/browse/TVIST1-788)
+  Updated case event sender and recipients to be simple text.
 
 ## [1.1.2] 2023-02-17